Se le dio funcion al boton modify cargando el form con todos los datos a modifcar del cliente atravez de un metodo que toma el id de la grilla

+++ BackEnd/Model/CustomerModel.cs
@@ -8,7 +8,7 @@ public class CustomerModel
     public string StreetCustomer { get; set; }
     public int StreetNumberCustomer { get; set; }
     public NeighborhoodModel Neighborhood { get; set; }
-    public int TelCustomer { get; set; }
+    public long TelCustomer { get; set; }
     public string MailCustomer { get;set; }
 
     public CustomerModel()
